services/dashdeid/requirements.txt
dash==2.11.1
dash-ag-grid==2.3.0
dash-bootstrap-components==1.4.2
dash-iconify==0.1.2
dash-mantine-components==0.12.1
diskcache==5.6.1
girder-client==3.1.22
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
numpy==1.24.4
pandas==2.0.3
pandocfilters==1.5.0
Pillow==9.4.0
plotly==5.15.0
psutil==5.9.5
pylibdmtx==0.1.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
requests==2.31.0
webcolors==1.13
girder==3.2.2
